+++ weblate/trans/models/unit.py
@@ -207,6 +207,9 @@ class UnitManager(models.Manager):
                 params['ignored']
             )
 
+        if 'lang' in params and params['lang']:
+            base = base.filter(translation__language__code=params['lang'])
+
         if not params['q']:
             return base